revised version of DieAndDiceSetTestHarness.java
an error was found in this tester which caused confusion.  In order to test the "setSides()" method of the Die.java class, I included a "getSides()" method in MY Die.java file, which was tested in four places in the DieAndDiceSetTestHarness.java file.  This situation was causing the test harness to produce compile errors.  Since this method was NOT REQUIRED for the assignment, and there was no template method provided, I have updated the test harness by commenting out the calls to the "getSides()" method.